refactor: override fetch method
