+++ cyberdrop_dl/crawlers/coomer.py
@@ -6,9 +6,8 @@ from aiolimiter import AsyncLimiter
 from yarl import URL
 
 from cyberdrop_dl.clients.errors import ScrapeError
-from cyberdrop_dl.crawlers.crawler import create_task_id
-from cyberdrop_dl.crawlers.kemono import KemonoCrawler
-from cyberdrop_dl.utils.utilities import error_handling_wrapper, remove_parts
+from cyberdrop_dl.crawlers.kemono import KemonoCrawler, KemonoPost
+from cyberdrop_dl.utils.utilities import error_handling_wrapper
 
 if TYPE_CHECKING:
     from cyberdrop_dl.managers.manager import Manager
@@ -29,24 +28,6 @@ class CoomerCrawler(KemonoCrawler):
         self.session_cookie = self.manager.config_manager.authentication_data.coomer.session
 
     """~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~~"""
-
-    @create_task_id
-    async def fetch(self, scrape_item: ScrapeItem) -> None:
-        """Determines where to send the scrape item based on the url."""
-        if "thumbnails" in scrape_item.url.parts:
-            scrape_item.url = remove_parts(scrape_item.url, "thumbnails")
-            return await self.handle_direct_link(scrape_item)
-        if "discord" in scrape_item.url.parts:
-            return await self.discord(scrape_item)
-        if "post" in scrape_item.url.parts:
-            return await self.post(scrape_item)
-        if scrape_item.url.name == "posts" and scrape_item.url.query.get("q"):
-            return await self.search(scrape_item)
-        if any(x in scrape_item.url.parts for x in self.services):
-            return await self.profile(scrape_item)
-        elif "favorites" in scrape_item.url.parts:
-            await self.favorites(scrape_item)
-        await self.handle_direct_link(scrape_item)
 
     @error_handling_wrapper
     async def favorites(self, scrape_item: ScrapeItem) -> None:
@@ -75,10 +56,15 @@ class CoomerCrawler(KemonoCrawler):
             new_scrape_item = scrape_item.create_child(url)
             self.manager.task_group.create_task(self.run(new_scrape_item))
 
-    @error_handling_wrapper
-    async def _handle_post_content(self, scrape_item: ScrapeItem, post: dict) -> None:
+    def _handle_post_content(self, scrape_item: ScrapeItem, post: KemonoPost) -> None:
         """Handles the content of a post."""
-        if "#ad" in post["content"] and self.manager.config_manager.settings_data.ignore_options.ignore_coomer_ads:
+        if "#ad" in post.content and self.manager.config_manager.settings_data.ignore_options.ignore_coomer_ads:
             return
 
-        super()._handle_post_content(scrape_item, post)  # type: ignore
+        return super()._handle_post_content(scrape_item, post)  # type: ignore
+
+    async def handle_direct_link(self, scrape_item: ScrapeItem) -> None:
+        """Handles a direct link."""
+        if "favorites" in scrape_item.url.parts:
+            return await self.favorites(scrape_item)
+        return await super().handle_direct_link(scrape_item)
